Make editCommand abstract
